Update trx4.ino
The "Range" setting adjusts the rotation limit of the servos, essentially setting the minimum and maximum positions the servos can move to. This is to protecting the servos preventing the servos from over-rotating, which can cause damage to the servo or mechanical linkages. It also allows for customizing suspension travel.  Allows the fine-tuning of the suspension travel to match the physical characteristics of the TRX4. This is intended to help ensure that the suspension behaves realistically and within the intended mechanical limits.